Add application framework
A very coarse implementation of a command-line-driven application.

The CLI interface has not been worked much on, the print outs look pathetic, but the program works and is runnable.

application.py needs to be rewritten. It is messy and contains logic that is best separated into several modules. In the end I wanted something runnable, from which to continue working. This purpose it fulfills.

Nothing is being commited to the database as is, but it runs on the database in memory. Need to polish things a bit more before database committing feels relevant.